Remove manager templates
I decided that a better option would be the use of groups.  When a
download that has restrictions on downloading multiple files in
parallel, they would be put in a group.  This group object will control
that sort of thing.  It allows a general class to deal with specific
issues of a service.